Flag analyses with ISubmitted and IVerified when out-of-stock
If the sample does only have one analysis and this analysis is
in out-of-stock status, the transitions pre-publish and publish
are not available for sample. Thus, becomes not possible to
generate results reports for samples of this kind.

These changes overcome this problem by flagging the analyses with
ISubmitted and IVerified interfaces.

Related: https://github.com/senaite/senaite.core/pull/2468